Refactor RacingGame Structure
