+++ src/main/java/exception/OutOfCarNameLengthException.java
@@ -1,3 +1,5 @@
+package exception;
+
 public class OutOfCarNameLengthException extends Exception {
 
     public OutOfCarNameLengthException(String message) {
